debugged navBarCtrl
since on my scope I had it set as '=' it was expecting me to pass
in a test variable when calling the directive.

for example <nav-bar test = "bacon"> </nav-bar>

Since I specifically designated a controller for this directive
everything that is put on scope in the controller will already be
on scope.